+++ tools/idea-plugin/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 - Migrate to IDEA code highlighting for tooltip snippets
+- Migrate to Jewel 0.34 and IntelliJ IDEA 2025.3.3
 
 ## 1.3.0 - 2026-02-28